[wip] extract MemoryCache to module
